Fix wrongly using fallback transition when selection gets lost
When the album data changes (e.g. video filter changes, photo library data changes)  while the editor is on-screen, the transition would revert to the fallback transition even though the edited video could still be present in the album.

This is due to the collection view clearing its selection on any data change. Keep track of the selection manually and restore it during the transitoin.