Fix editor sometimes not animating to correct cell
In some cases, the navigation controller collapses its large navigation bar during the pop transition.

(To reproduce: This happened mostly for albums with more than one but less than a screen height full of rows. It happens both during the custom and the system slide to back transition.)

For the non-interactive pop transition (back button), the target frame of the original position would get reported and the transition would erroneously animate to there. I did not find a way to get the final position of the cell after the bar has collapsed. For the interactive transition on the other hand, the final frame would be correctly reported and this issue was not present.

The current solution is a workaround in that we force collapse the navigation bar for *every* transition for the duration of the transition.