Don't change gridMode in willDisplayCell
There are a couple of non-conclusive crash reports hinting to willDisplayCell after changing the grid content mode.

I'm not sure what the cause is this so this is a stab in the dark: Avoid changing the content mode in willDisplayCell. Instead use a workaraound of animating the visible cells and using reloadData for enqueue off-screen cells after the animation finishes.