Snap playback times to the video track's timescale
Fixes an issue when stepping frame by frame in some videos.

When using `step(byCount:)`, `AVPlayer` steps to the desired sample but reports the playback time with a timescale of `NSEC_PER_SEC`. Due to the different timescale, the playback time is smaller than the desired sample time by a miniscule amount. Since it is smaller, `sampleTiming(for:)` returns the preceding, but unexpected, sample in that case.

By first snapping the playback time to the track's timescale, we can avoid this issue.

However, I am not entirely clear if this introduces other rounding issues for different videos or sample timings. We'll see.