Remove incremental photo updates
Declare bankruptcy and remove incremental updates in favor of simple reloading.

PhotosKit and incremental updates have lots of problems. First, incremental updates are notoriously fragile. The reported change indexes often lead to crashes in `performBatchUpdates` without any indication as to why.

Next, incremental updates tie us to using the fixed sort order of the underlying fetch request. This is in conjuction with another limitation of PhotosKit: Smart album fetch requests return their result in "reverse order", i.e. the oldest photo comes first. This makes sense for Photos.app where the newest items are at the bottom but it is hard to adapt for collection views where the newest items should be at the top. Specifying a sort descriptor by creation date does not fix this because the Recents photo album is not sorted by creation date but by "added date". Well, here's the next limitation: PhotosKit does not support a sort descriptor for this order.

The result: Incremental updates and sort order  by "added date" but with the newest one at the top is simply not possible in a seamless experience. Which is ridiculous in my opinion because that is **exactly** what the user expects.

Possible solutions:

1) Ditch fetch requests, map its assets to an array and sort it however you want, maybe with custom diffing (e.g. diffable data sources). Works okaysh with up to 10k items but after that, the performance degrades fast. Photo libraries can be huge (100k+ items not an exception) so straying off of the batching behaviour of fetch results is not recommended by Apple.

2) Manually reverse indexes of the fetch request to get the desired order. Then try to map incremental update indexes to those reversed indexes. Well, incremental update indexes are already fragile and often wrong as they are. Attempting this would lead to an implosion.

3) Ditch incremental updates.

I'm going with 3. Just use reloadData. Then, we can get the correct "added date" sort order simply by reversing the indexes.

Unless I'm missing something reaaally obvious, I think it incremental changes and PhotosKit in general have fundamental flaws in their design.

Rant end. 🥵