Add custom collection view reload animation
Since we don't do incremental updates anymore, we use plain `reloadData`. It looks boring though and doesn't give a lot of visual indication to the user when something changes. `reloadSections` looks better but it degrades above 10k items in the collection view.

This simple animation works with any number of items. Hopefully, there are no side-effects in using a snapshot.